HZN-1185: Move osgiscript to common folder
