+++ checkstyle/src/main/resources/license-suppressions.xml
@@ -10,9 +10,9 @@
     <suppress files="org/opennms/netmgt/snmp/snmp4j/Integer32IgnoreTooManyBytes.java" checks=".*"/>
     <suppress files="org/opennms/core/soa/support/OnmsOSGiBridgeActivator.java" checks=".*"/>
     <suppress files="org/opennms/vaadin/extender/.*" checks=".*"/>
-    <suppress files="org/opennms/features/topology/plugins/topo/graphml/internal/scripting/.*" checks=".*"/>
     <suppress files="org/opennms/poller/remote/Mac.*" checks=".*"/>
     <suppress files="org/opennms/container/web/.*" checks=".*"/>
+    <suppress files="org/opennms/features/osgi/script/.*" checks=".*"/>
     <!-- BSD style -->
     <suppress files="org/opennms/netmgt/syslogd/SyslogClient.java" checks=".*"/>
     <suppress files="org/opennms/plugins/com/impossibl/postgres/jdbc/.*" checks=".*"/>
