+++ features/osgi-jsr223/src/main/java/org/opennms/features/osgi/script/OSGiScriptEngineFactory.java
@@ -14,7 +14,7 @@
  *   limitations under the License.
  *
  */
-package org.opennms.features.topology.plugins.topo.graphml.internal.scripting;
+package org.opennms.features.osgi.script;
 
 import java.util.List;
