+++ features/telemetry/adapters/collection/pom.xml
@@ -53,5 +53,10 @@
       <artifactId>org.osgi.core</artifactId>
       <scope>provided</scope>
     </dependency>
+    <dependency>
+      <groupId>org.opennms.features</groupId>
+      <artifactId>org.opennms.features.osgi-jsr223</artifactId>
+      <version>${project.version}</version>
+    </dependency>
   </dependencies>
 </project>
