--- features/topology-map/plugins/org.opennms.features.topology.plugins.topo.graphml/src/main/java/org/opennms/features/topology/plugins/topo/graphml/internal/scripting/.readme
@@ -1,1 +1,0 @@
-These classes have been copied from the apache felix project, as they are not accessible or available as a separate jar (bundle).
